requirements.txt
boolean.py
cachetools
GitPython
inflect
ipython
jsonschema
jupyter
matplotlib
msgpack
nltk
numpy
pandas
polars
pyarrow
pytest
python_Levenshtein
scikit_learn
scikit-spatial
scipy
statsmodels
tabulate
TatSu
torch
tqdm
viztracer
wandb
